Restructure main control flow
The former main() function is now part of the Renamer class, which
keeps the formerly global state for dry runs ("simulate"). The new
main() function does basically everything that used to be part of the
part after the __main__ guard.

This makes it possible to call main() with what would be command line
parameters, which is useful for tests without starting a new
interpreter.

Closes #18.