Replace StderrLogger with the logging module
The logging module provides functions for different log levels. I've
switched to using their formatting instead of format strings because I
assume in that case the formatting will only be done if needed, which
is a minor performance improvement.

The new --quiet and --debug options allow changing the log level for
the output to WARNING or DEBUG, respectively. With --quiet only errors
(missing files, files without a usable date source) will be logged.

Note that only the first call to logging.basicConfig() takes effect,
so we can conveniently override the configuration for tests.