Fix license field in pyproject.toml
I noticed https://github.com/wntrblm/nox/pull/619 and it turns out the
issue affects "pip show" as well: If the license field is set to a
file, the whole content of the file is quoted in the pip output, which
is unwieldy. Specify a short identifier instead, the COPYING file is
still available.